move homeview, update styles & text, clean view logic
