--- retrospective/forms.py
@@ -1,9 +1,0 @@
-from django.forms import ModelForm
-
-from .models import RetrospectiveEvent
-
-
-class RetrospectiveEventForm(ModelForm):
-    class Meta:
-        model = RetrospectiveEvent
-        fields = ["date", "certainty"]
