+++ retrospective/views.py
@@ -1,40 +1,22 @@
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import render, redirect
-from django.views.generic import TemplateView, FormView
+from django.views.generic import CreateView
 from django.http import HttpResponse
 from django.contrib import messages
-from openhumans.models import OpenHumansMember
 
-from .forms import RetrospectiveEventForm
 from .tasks import analyze_event
 from .models import RetrospectiveEventAnalysis, RetrospectiveEvent
-from .helpers import identify_missing_sources
-from django.contrib.auth.decorators import login_required
 
 
-class HomeView(TemplateView):
-    template_name = "retrospective/home.html"
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-        context["openhumans_login_url"] = OpenHumansMember.get_auth_url()
-        if self.request.user.is_authenticated:
-            missing_sources = identify_missing_sources(
-                self.request.user.openhumansmember
-            )
-            context["missing_sources"] = missing_sources
-        return context
-
-
-class AddRetrospectiveEventView(LoginRequiredMixin, FormView):
+class AddRetrospectiveEventView(LoginRequiredMixin, CreateView):
+    model = RetrospectiveEvent
+    fields = ["date", "certainty"]
     template_name = "retrospective/addevent.html"
-    form_class = RetrospectiveEventForm
     success_url = "/"
     login_url = "/"
 
     def form_valid(self, form):
-        # This method is called when valid form data has been POSTed.
-        # It should return an HttpResponse.
         form.instance.member = self.request.user.openhumansmember
         event = form.save()
         print("ANALYZING EVENT {}".format(event.id))
